weekend work

+++ database/migrations/2015_04_02_121533_create_mount_table.php
@@ -11,19 +11,30 @@ class CreateMountTable extends Migration
      */
     public function up()
     {
-        \Schema::create(
-            'mount_t',
+        if ( !\Schema::hasTable( 'mount_t' ) )
+        {
+            \Schema::create(
+                'mount_t',
+                function ( Blueprint $table )
+                {
+                    $table->increments( 'id' );
+                    $table->integer( 'mount_type_nbr' )->default( 0 );
+                    $table->string( 'mount_id_text', 64 )->unique();
+                    $table->integer( 'owner_id' )->default( 0 );
+                    $table->string( 'root_path_text', 128 )->nullable();
+                    $table->mediumText( 'config_text' )->nullable();
+                    $table->dateTime( 'last_mount_date' )->nullable();
+                    $table->dateTime( 'create_date' );
+                    $table->timestamp( 'lmod_date' )->default( \DB::raw( 'CURRENT_TIMESTAMP' ) );
+                }
+            );
+        }
+
+        !\Schema::hasColumn( 'server_t', 'mount_id' ) && \Schema::table(
+            'server_t',
             function ( Blueprint $table )
             {
-                $table->increments( 'id' );
-                $table->integer( 'mount_type_nbr' )->default( 0 );
-                $table->string( 'mount_id_text', 64 )->unique();
-                $table->integer( 'owner_id' )->default( 0 );
-                $table->string( 'root_path_text', 128 )->nullable();
-                $table->mediumText( 'config_text' )->nullable();
-                $table->dateTime( 'last_mount_date' )->nullable();
-                $table->dateTime( 'create_date' );
-                $table->timestamp( 'lmod_date' )->default( \DB::raw( 'CURRENT_TIMESTAMP' ) );
+                $table->integer( 'mount_id' )->nullable();
             }
         );
     }
@@ -35,7 +46,10 @@ class CreateMountTable extends Migration
      */
     public function down()
     {
-        Schema::drop( 'mount_t' );
+        if ( \Schema::hasTable( 'mount_t' ) )
+        {
+            Schema::drop( 'mount_t' );
+        }
     }
 
 }
